Create micenewdata.txt
So this will show the crazy data were looking for(i.e. so we have to to scroll). For some reason my one that I input all the generation info is not working with the visualization, so I think I just have a syntax error or I have a wrong litter/generation. Will fix ASAP, but this will help with vis!